# docs/swagger.yaml
definitions:
  dto.UserDto:
    properties:
      age:
        type: integer
      gender:
        type: string
      name:
        type: string
    type: object
  entity.User:
    properties:
      age:
        type: integer
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /users:
    get:
      description: retrieve all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
        "204":
          description: No Content
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Find all users
      tags:
      - users
    post:
      description: create a user
      parameters:
      - description: Payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create user
      tags:
      - users
  /users/{id}:
    delete:
      description: delete a user by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete user by ID
      tags:
      - users
    get:
      description: retrieve a user by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Find user by ID
      tags:
      - users
    put:
      description: update a user by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update user by ID
      tags:
      - users
swagger: "2.0"